Added org.json, Cleaned Telemetry, Added waypoints, Moved Controllers
Added org.json due to it being more familiar. I plan to make a kotlin application that allows for easy updating and creation of waypoints. 

Telemetry no longer has html tags since it is not converted to html on the dashboard or driver-station.

Added waypoint values based on a picture of the field. They have not been measured on a physical field.

Controllers are now initialized and updated in RobotHardware, rather than in the different opmodes.

Dashboard now displays telemetry and has a few variables that are configurable.