Fixed inability to change launcher speed while driving, Commented out deadwheels
Commented out dead wheel motors since they have yet to be installed and caused confusion about the missing motors when starting the robot.

Fixed the issue of being unable to change the launcher's speed during driver period since it was constantly being reset to a value depending on if 'powershotMode' was enabled.

Changed the next state after shooting to be 'Park' in order to test autonomous.