Fixed endpoint to be START_WALL instead of PARK
I was getting the start and end of the auto trajectory 
confused with the start and end of the return from 
auto trajectory.
In this case, the PARK waypoint is NOT where we want to 
park.  Instead, we want to end our trajectory by 
parking at START_WALL, where autonomous starts.