Draft full states
Vim states for powershot routine

Spent an hour figuring out how to use vim, 
then pressed the correct 12 keystrokes in the 
correct order, and this commit was generated in a single 
blistering fast step.  Magic.

The states just run a trajectory and go to the next state.
They need more excitement in their lives.