Setup drive test in manual to evaluate start position angle error
Concept:
Run Manual with the robot in the START_CENTER position.
Press LB + X to drive 96" forward, with a y of -18"
After the drive, measure the y position of the robot.
If there is y error, this may be caused by an error 
in the starting heading.

At present, this position is assumed to be square with a 
heading of 180 degrees.
However, this may need an angular correction:
angleCorrection = atan(yError / 96)