LauncherControl.calibrateFeedForward() drafted
When an instance of LauncherControl is created, it can be 
initialized in one of 3 states, each of which assigns 
initial PIDF paremters.
Integral: uses original PIDF(10,3,0,0)
FeedForwardOpenLoop and FeedForwardClosedLoop both 
initialize with PIDF(40,0,0,15),
however only FeedForwardClosedLoop allows the 
calibrateFeedForward() method to have any effect.

calibrateFeedForward(targetVelocityTicksPerSecond)
evaluates whether the steady state velocity is close enough 
to the target, then adjusts the F parameter if required.
- Selecting the correct F increment is a challenge, and 
I'm not certain that the included procedure is the best way.
A dumb system for just bumping by 0.3 might actually 
work if this doesn't.