Bound LauchControl F parameter increment
Calculating the F parameter increment is one of the main 
challenges of calibration, but enforcing upper and lower 
bounds helps keep things reasonable. 

On the low end, if the system has determined that there is 
enough error that calibration is required, then increments 
below a certain size won't do enough.
On the high end, giant jumps to F are likely to cause instability.