Added ListMath.getLinearRegression(time,data) with tests
Linear Regression tools provided and demonstrated in test, 
pulled from the Apache common math3 library.
Slope, Intercept, MeanSquaredError and more are all 
easily accessible.  Processing time has not be tested.

A potential use is to replace the isSteadyState test with 
a linear regression based test so shots can be fired even 
when not at steady state velocity IF the error is low enough 
and if the error is being reduced without overshoot.