Distance Utility, getPositionFromRange()
Setup system for caching range sensor output such that 
it is never called more than once per cycle, and only when it is used.

getPositionFromRange( positionEstimate ) takes in the two 
range sensor readings and calculates a new pose for the 
robot.  Only operates if pointed within a few degrees of 180
 heading, and additional guard conditions could be added.